backend/pyproject.toml
[tool.poetry]
name = "traffic-monitor-fastapi"
version = "0.1.0"
description = ""
authors = ["Timspizza <peisen.jiang2001@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.6"
pypcap = "^1.3.0"
scapy = "^2.6.1"
uvicorn = "^0.34.0"
netifaces = "^0.11.0"
py-spy = "^0.4.0"
python-dotenv = "^1.0.1"
pymongo = "^4.10.1"
geoip2 = "^4.8.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.20"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
